Change import syntax and rename files to PascalCase
